Fix motion blur
Camera's times weren't set

+++ RaytracingBooksCode/ray.h
@@ -18,7 +18,7 @@ struct ray
 
     vec3 orig;
     vec3 dir;
-    double tm;
+    float tm;
 };
 
 #endif
